+++ test/foundry/StandardTransactions.t.sol
@@ -187,7 +187,7 @@ contract StandardTransactionsTest is ProtocolBase {
                 0, // subsetNonce
                 0, // strategyId (Standard sale for fixed price)
                 0, // assetType ERC721,
-                uint112(i), // orderNonce
+                i, // orderNonce
                 address(mockERC721),
                 address(0), // ETH,
                 makerUser,
@@ -239,7 +239,7 @@ contract StandardTransactionsTest is ProtocolBase {
             // Taker user has received the asset
             assertEq(mockERC721.ownerOf(i), takerUser);
             // Verify the nonce is marked as executed
-            assertEq(looksRareProtocol.userOrderNonce(makerUser, uint112(i)), MAGIC_VALUE_NONCE_EXECUTED);
+            assertEq(looksRareProtocol.userOrderNonce(makerUser, i), MAGIC_VALUE_NONCE_EXECUTED);
         }
         // Taker bid user pays the whole price
         assertEq(address(takerUser).balance, _initialETHBalanceUser - (numberPurchases * price));
@@ -273,7 +273,7 @@ contract StandardTransactionsTest is ProtocolBase {
                 0, // subsetNonce
                 0, // strategyId (Standard sale for fixed price)
                 0, // assetType ERC721,
-                uint112(i), // orderNonce
+                i, // orderNonce
                 address(mockERC721),
                 address(0), // ETH,
                 makerUser,
@@ -354,13 +354,13 @@ contract StandardTransactionsTest is ProtocolBase {
             // Taker user has received the first two assets
             assertEq(mockERC721.ownerOf(i), takerUser);
             // Verify the first two nonces are marked as executed
-            assertEq(looksRareProtocol.userOrderNonce(makerUser, uint112(i)), MAGIC_VALUE_NONCE_EXECUTED);
+            assertEq(looksRareProtocol.userOrderNonce(makerUser, i), MAGIC_VALUE_NONCE_EXECUTED);
         }
 
         // Taker user has not received the asset
         assertEq(mockERC721.ownerOf(faultyTokenId), randomUser);
         // Verify the nonce is NOT marked as executed
-        assertEq(looksRareProtocol.userOrderNonce(makerUser, uint112(faultyTokenId)), bytes32(0));
+        assertEq(looksRareProtocol.userOrderNonce(makerUser, faultyTokenId), bytes32(0));
         // Taker bid user pays the whole price
         assertEq(address(takerUser).balance, _initialETHBalanceUser - 1 - ((numberPurchases - 1) * price));
         // Maker ask user receives 98% of the whole price (2% protocol)
